requirements.txt
pyshorteners
python-whois
mysql-connector-python